refactor(rspack_core): move ModuleGraphConnection to its own file
