+++ crates/rspack_core/src/module_graph/mod.rs
@@ -1,64 +1,17 @@
 use std::borrow::Cow;
-use std::cmp::PartialEq;
 use std::collections::hash_map::Entry;
-use std::hash::{Hash, Hasher};
 use std::sync::atomic::{AtomicUsize, Ordering};
 
 use rspack_error::{internal_error, Result};
 use rspack_identifier::IdentifierMap;
 use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
 
+mod connection;
+pub use connection::ModuleGraphConnection;
+
 use crate::{
   BoxModule, BoxModuleDependency, DependencyId, Module, ModuleGraphModule, ModuleIdentifier,
 };
-
-// FIXME: placing this as global id is not acceptable, move it to somewhere else later
-static NEXT_MODULE_GRAPH_CONNECTION_ID: AtomicUsize = AtomicUsize::new(1);
-
-#[derive(Debug, Clone, Eq)]
-pub struct ModuleGraphConnection {
-  /// The referencing module identifier
-  pub original_module_identifier: Option<ModuleIdentifier>,
-  /// The referenced module identifier
-  pub module_identifier: ModuleIdentifier,
-  /// The referencing dependency id
-  pub dependency_id: usize,
-
-  /// The unique id of this connection
-  pub id: usize,
-}
-
-impl Hash for ModuleGraphConnection {
-  fn hash<H: Hasher>(&self, state: &mut H) {
-    self.original_module_identifier.hash(state);
-    self.module_identifier.hash(state);
-    self.dependency_id.hash(state);
-  }
-}
-
-impl PartialEq for ModuleGraphConnection {
-  fn eq(&self, other: &Self) -> bool {
-    self.original_module_identifier == other.original_module_identifier
-      && self.module_identifier == other.module_identifier
-      && self.dependency_id == other.dependency_id
-  }
-}
-
-impl ModuleGraphConnection {
-  pub fn new(
-    original_module_identifier: Option<ModuleIdentifier>,
-    dependency_id: usize,
-    module_identifier: ModuleIdentifier,
-  ) -> Self {
-    Self {
-      original_module_identifier,
-      module_identifier,
-      dependency_id,
-
-      id: NEXT_MODULE_GRAPH_CONNECTION_ID.fetch_add(1, Ordering::Relaxed),
-    }
-  }
-}
 
 #[derive(Debug, Default)]
 pub struct ModuleGraph {
